+++ jetty-core/jetty-http3/jetty-http3-qpack/src/main/java/org/eclipse/jetty/http3/qpack/internal/metadata/Http3Fields.java
@@ -121,12 +121,6 @@ public class Http3Fields implements HttpFields
     }
 
     @Override
-    public HttpFields takeAsImmutable()
-    {
-        return asImmutable();
-    }
-
-    @Override
     public HttpField getField(int index)
     {
         return stream().skip(index).findFirst().orElse(null);
