+++ jetty-core/jetty-server/src/main/java/org/eclipse/jetty/server/handler/gzip/GzipRequest.java
@@ -107,8 +107,7 @@ public class GzipRequest extends Request.Wrapper
                 default -> newFields.add(field);
             }
         }
-        fields = newFields.takeAsImmutable();
-        return fields;
+        return newFields.asImmutable();
     }
 
     @Override
