+++ jetty-ee9/jetty-ee9-tests/jetty-ee9-test-integration/src/test/java/org/eclipse/jetty/ee9/test/rfcs/RFC2616BaseTest.java
@@ -355,12 +355,12 @@ public abstract class RFC2616BaseTest
 
         fields.put("Q", "bbb;q=0.5,aaa,ccc;q=0.002,d;q=0,e;q=0.0001,ddd;q=0.001,aa2,abb;q=0.7");
         List<String> list = fields.getQualityCSV("Q");
-        assertEquals("aaa", HttpField.valueParameters(list.get(0).toString(), null), "Quality parameters");
-        assertEquals("aa2", HttpField.valueParameters(list.get(1).toString(), null), "Quality parameters");
-        assertEquals("abb", HttpField.valueParameters(list.get(2).toString(), null), "Quality parameters");
-        assertEquals("bbb", HttpField.valueParameters(list.get(3).toString(), null), "Quality parameters");
-        assertEquals("ccc", HttpField.valueParameters(list.get(4).toString(), null), "Quality parameters");
-        assertEquals("ddd", HttpField.valueParameters(list.get(5).toString(), null), "Quality parameters");
+        assertEquals("aaa", HttpField.getValueParameters(list.get(0), null), "Quality parameters");
+        assertEquals("aa2", HttpField.getValueParameters(list.get(1), null), "Quality parameters");
+        assertEquals("abb", HttpField.getValueParameters(list.get(2), null), "Quality parameters");
+        assertEquals("bbb", HttpField.getValueParameters(list.get(3), null), "Quality parameters");
+        assertEquals("ccc", HttpField.getValueParameters(list.get(4), null), "Quality parameters");
+        assertEquals("ddd", HttpField.getValueParameters(list.get(5), null), "Quality parameters");
     }
 
     /**
